Refactor code like Spring Boot project architect
